+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 5.0.0
+* Added `hbitdepth`, `miniswhite` options to `heif` command.
+* Added `lineArt` option to `trim` command.
+* Updated `sharp` dependency (requiring dependency to bump Node to 18.17+).
+
 ## 4.2.0 (November 25, 2023)
 * Exit process when there are no input files ([#83](https://github.com/vseventer/sharp-cli/issues/83)).
 * Added `unflatten` command.